+++ R/daily_response.R
@@ -102,7 +102,7 @@
 #' the highest calculated metric and is obtained by the $plot_extreme output, is the
 #' same for all three reference windows.
 #'
-#' @return a list with 14 elements:
+#' @return a list with 15 elements:
 #' \tabular{rll}{
 #'  1 \tab $calculations   \tab a matrix with calculated metrics\cr
 #'  2 \tab $method \tab the character string of a method \cr
@@ -117,7 +117,8 @@
 #'  11 \tab $plot_extreme    \tab ggplot2 object: line plot of a row with the highest value in a matrix of calculated metrics\cr
 #'  12 \tab $plot_specific    \tab ggplot2 object: line plot of a row with a selected window width in a matrix of calculated metrics\cr
 #'  13 \tab $PCA_output    \tab princomp object: the result output of the PCA analysis\cr
-#'  14 \tab $type    \tab the character string describing type of analysis: daily or monthly
+#'  14 \tab $type    \tab the character string describing type of analysis: daily or monthly\cr
+#'  15 \tab $reference_window \tab character string, which referece window was used for calculations
 #'}
 #'
 #' @export
@@ -141,8 +142,11 @@
 #'                                      row_names_subset = TRUE, remove_insignificant = TRUE,
 #'                                      alpha = 0.05, aggregate_function = 'mean',
 #'                                      reference_window = "end")
+#' class(example_daily_response)
 #' example_daily_response$plot_extreme
 #' example_daily_response$plot_heatmap
+#' summary.ds(example_daily_response)
+#' summary(example_daily_response)
 #'
 #' # 2 Example for past and present
 #' example_MVA_past <- daily_response(response = data_MVA, env_data = LJ_daily_temperatures,
@@ -1850,7 +1854,8 @@ for (m in 1:length(empty_list_datasets)){
                          plot_extreme = plot_extremeA,
                          plot_specific = plot_specificA,
                          PCA_output = PCA_result,
-                         type = "daily")
+                         type = "daily",
+                         reference_window = reference_window)
     }
 
     if (method == "cor"){
@@ -1864,8 +1869,11 @@ for (m in 1:length(empty_list_datasets)){
                          plot_extreme = plot_extremeA,
                          plot_specific = plot_specificA,
                          PCA_output = PCA_result,
-                         type = "daily")
+                         type = "daily",
+                         reference_window = reference_window)
     }
+
+    class(final_list) <- 'ds'
 
   return(final_list)
 }